+++ lib/utils/widgets/content_card.dart
