+++ lib/utils/widgets/terms_card.dart
