+++ lib/views/landing_page.dart
@@ -1,10 +1,10 @@
-import 'package:diva_roulette/views/desktop/accountCreatePageDesktop.dart';
-import 'package:diva_roulette/views/tablet/accountCreationPageTablet.dart';
-import 'package:diva_roulette/views/mobile/accountCreationPageMobile.dart';
+import 'package:diva_roulette/views/desktop/landing_page_desktop.dart';
+import 'package:diva_roulette/views/mobile/landing_page_mobile.dart';
+import 'package:diva_roulette/views/tablet/landing_page_tablet.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_builder/responsive_builder.dart';
 
-class AccountCreationPage extends StatelessWidget {
+class LandingPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -12,12 +12,12 @@ class AccountCreationPage extends StatelessWidget {
         child: Scrollbar(
           child: ResponsiveBuilder(builder: (context, sizingInformation) {
             if (sizingInformation.deviceScreenType == DeviceScreenType.mobile) {
-              return AccountCreationPageMobile();
+              return LandingPageMobile();
             } else if (sizingInformation.deviceScreenType ==
                 DeviceScreenType.tablet) {
-              return AccountCreationPageTablet();
+              return LandingPageTablet();
             } else {
-              return AccountCreationPageDesktop();
+              return LandingPageDesktop();
             }
           }),
         ),
