+++ lib/views/login_page.dart
@@ -1,10 +1,10 @@
-import 'package:diva_roulette/views/desktop/landingPageDesktop.dart';
-import 'package:diva_roulette/views/mobile/landingPageMobile.dart';
-import 'package:diva_roulette/views/tablet/landingPageTablet.dart';
+import 'package:diva_roulette/views/desktop/login_page_desktop.dart';
+import 'package:diva_roulette/views/mobile/login_page_mobile.dart';
+import 'package:diva_roulette/views/tablet/login_page_tablet.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_builder/responsive_builder.dart';
 
-class LandingPage extends StatelessWidget {
+class LoginPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -12,12 +12,12 @@ class LandingPage extends StatelessWidget {
         child: Scrollbar(
           child: ResponsiveBuilder(builder: (context, sizingInformation) {
             if (sizingInformation.deviceScreenType == DeviceScreenType.mobile) {
-              return LandingPageMobile();
+              return LoginPageMobile();
             } else if (sizingInformation.deviceScreenType ==
                 DeviceScreenType.tablet) {
-              return LandingPageTablet();
+              return LoginPageTablet();
             } else {
-              return LandingPageDesktop();
+              return LoginPageDesktop();
             }
           }),
         ),
