+++ lib/views/mobile/account_creation_page_mobile.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:auto_size_text/auto_size_text.dart';
-import 'package:diva_roulette/views/mobile/verifyDialogMobile.dart';
+import 'package:diva_roulette/views/mobile/verify_dialog_mobile.dart';
 import 'package:flutter/material.dart';
 import 'package:getwidget/getwidget.dart';
 import 'package:ndialog/ndialog.dart';
