+++ lib/views/tablet/verify_dialog_tablet.dart
